refactor: unify SQL prefixes to 'build' and consolidate functions into LibsqlQueryBuilder container
