+++ src/client/providers/libsql/libsql-search-index.ts
@@ -5,7 +5,7 @@ import type {
   SearchResponse,
   SearchResult,
 } from "#/client/search-index/search-index-interface.ts";
-import { buildSearchQuery } from "./statements.ts";
+import { LibsqlQueryBuilder } from "./libsql-query-builder.ts";
 
 import type { EmbeddingService } from "#/client/search-index/embedding-service/mod.ts";
 
@@ -47,7 +47,7 @@ export class LibsqlSearchIndex implements SearchIndexInterface {
       );
     }
 
-    const { sql, args } = buildSearchQuery(request, {
+    const { sql, args } = LibsqlQueryBuilder.buildSearchQuery(request, {
       vectorJson,
       limit: this.options.limit ?? 100,
     });
