+++ src/client/providers/libsql/sync-libsql.test.ts
@@ -4,24 +4,17 @@ import { DataFactory } from "n3";
 import { RecursiveCharacterTextSplitter } from "@langchain/textsplitters";
 import { syncLibsql } from "./sync-libsql.ts";
 import { FakeEmbeddingService } from "#/client/search-index/embedding-service/mod.ts";
-import {
-  makeLibsqlChunksFtsTable,
-  makeLibsqlChunksIndex,
-  makeLibsqlChunksQuadIdIndex,
-  makeLibsqlChunksTable,
-  makeLibsqlChunksTriggers,
-  makeLibsqlQuadsTable,
-} from "./statements.ts";
+import { LibsqlQueryBuilder } from "./libsql-query-builder.ts";
 
 const { quad, namedNode, literal } = DataFactory;
 
 async function setupSchema(client: ReturnType<typeof createClient>) {
-  await client.execute(makeLibsqlQuadsTable()); // <--- Demand Quads Table exist
-  await client.execute(makeLibsqlChunksTable());
-  await client.execute(makeLibsqlChunksQuadIdIndex());
-  await client.execute(makeLibsqlChunksFtsTable());
-  await client.execute(makeLibsqlChunksIndex());
-  for (const triggerSql of makeLibsqlChunksTriggers()) {
+  await client.execute(LibsqlQueryBuilder.buildLibsqlQuadsTable()); // <--- Demand Quads Table exist
+  await client.execute(LibsqlQueryBuilder.buildLibsqlChunksTable());
+  await client.execute(LibsqlQueryBuilder.buildLibsqlChunksQuadIdIndex());
+  await client.execute(LibsqlQueryBuilder.buildLibsqlChunksFtsTable());
+  await client.execute(LibsqlQueryBuilder.buildLibsqlChunksIndex());
+  for (const triggerSql of LibsqlQueryBuilder.buildLibsqlChunksTriggers()) {
     await client.execute(triggerSql);
   }
 }
